on_keyboard_event hook
